requirements.txt
# Main
geocoder
httpx
orjson
tzlocal

# Development
camoufox
fastapi
rich
ruff
uvicorn

# Discord bot
audioop-lts
py-cord
python-dotenv
rich